pyproject.toml
[project]
name = "jeelink2mqtt"
dynamic = ["version"]
description = ""
authors = [{ name = "Alexander Herrmann"}]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
  "Development Status :: 4 - Beta",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Natural Language :: English",
  "Operating System :: OS Independent",
]
dependencies = ["click", "paho-mqtt", "pyserial-asyncio", "expandvars", "psutil"]

[project.scripts]
jeelink2mqtt-cli = "jeelink2mqtt.jeelink2mqtt:main"

[project.optional-dependencies]
#test = ["pytest", "pytest-cov"]
test = ["pytest"]
dev = ["pre-commit"]

[project.urls]
Homepage = "https://github.com/gandy92/jeelink2mqtt"

[tool.setuptools]
packages = ["jeelink2mqtt"]

[tool.setuptools.dynamic]
version = { attr = "jeelink2mqtt.__version__" }

[tool.pytest.ini_options]
#addopts = "--cov=jeelink2mqtt --cov-report html"
#addopts = "--cov=jeelink2mqtt"

#[tool.setuptools.package-data]
## https://til.simonwillison.net/python/pyproject#user-content-package-data
#demo_package = ["static/*", "templates/*"]

[build-system]
# this is actually default, but doesn't hurt either
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
